# docs/swagger.yaml
basePath: /
definitions:
  handler.BatchResponse:
    properties:
      cancelled:
        type: integer
      created_at:
        type: string
      failed:
        type: integer
      id:
        type: string
      pending:
        type: integer
      processing:
        type: integer
      queued:
        type: integer
      sent:
        type: integer
      total:
        type: integer
      updated_at:
        type: string
    type: object
  handler.CreateBatchRequest:
    properties:
      notifications:
        items:
          $ref: '#/definitions/handler.CreateRequest'
        maxItems: 1000
        minItems: 1
        type: array
    required:
    - notifications
    type: object
  handler.CreateBatchResponse:
    properties:
      batch_id:
        type: string
      count:
        type: integer
    type: object
  handler.CreateRequest:
    properties:
      channel:
        enum:
        - sms
        - email
        - push
        example: sms
        type: string
      content:
        example: Your verification code is 123456
        maxLength: 1600
        type: string
      idempotency_key:
        type: string
      priority:
        enum:
        - high
        - normal
        - low
        example: normal
        type: string
      recipient:
        type: string
      scheduled_at:
        example: "2026-01-01T12:00:00Z"
        type: string
    required:
    - channel
    - content
    - recipient
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.ListResponse:
    properties:
      notifications:
        items:
          $ref: '#/definitions/handler.NotificationResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  handler.NotificationResponse:
    properties:
      batch_id:
        type: string
      channel:
        type: string
      content:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      failed_at:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      priority:
        type: string
      provider_id:
        type: string
      recipient:
        type: string
      retry_count:
        type: integer
      scheduled_at:
        type: string
      sent_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Scalable notification delivery service for SMS, Email, and Push channels
    with priority queuing, retry logic, and observability.
  title: Notification Service API
  version: "1.0"
paths:
  /api/v1/batches/{batchID}:
    get:
      parameters:
      - description: Batch ID
        in: path
        name: batchID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BatchResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get batch statistics
      tags:
      - batches
  /api/v1/notifications:
    get:
      parameters:
      - description: Filter by status (pending,queued,processing,sent,failed,cancelled)
        in: query
        name: status
        type: string
      - description: Filter by channel (sms,email,push)
        in: query
        name: channel
        type: string
      - description: Filter by batch ID
        in: query
        name: batch_id
        type: string
      - description: Filter from date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: Filter to date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List notifications
      tags:
      - notifications
    post:
      consumes:
      - application/json
      parameters:
      - description: Notification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a notification
      tags:
      - notifications
  /api/v1/notifications/{id}:
    delete:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Cancel a pending notification
      tags:
      - notifications
    get:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.NotificationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get notification by ID
      tags:
      - notifications
  /api/v1/notifications/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Batch payload (max 1000)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateBatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateBatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a batch of notifications
      tags:
      - notifications
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
swagger: "2.0"